requirements.txt
pandas==1.1.2
prompt_toolkit==1.0.14
regex==2020.11.13
PyInquirer==1.0.3
selenium==3.141.0
pyyaml==5.4.1